Fixes #1186: Add AccessibilityActions to quick action sheet's handle
Adding the 'click' action allows the user to activate the button with a
default action (double tap in TalkBack). This is instead of
conditionally setting up a click listener if TalkBack is enabled. This
is a more generalized solution that accomodates other ATs besides
TalkBack.

Adding an expand or collapse action also tells TalkBack what the current
state of the sheet is, and reports it to the user. So they will
initially hear "collapsed, quick actions, button". When they double tap,
they will hear "expanded". Adding those actions also allows TalkBack and
other AT users to explicitly call those actions on that element with a
menu or a gesture binding.