For #1525: Make BrowserPerformanceTestActivity pass through to IntentReceiverActivity.
Previously, this passed through to HomeActivity, but Intent arguments
were dropped.  That means that actions (`-a
android.intent.action.VIEW`) and URLs (`-d data:...`) were dropped,
which means that Gecko itself wasn't started (and wasn't started with
a session owned by Fenix itself).

This simply passes through the intent to its final destination.

Along the way, I also enabled remote debugging for these test
configurations; it's probably not necessary, but eventually I want all
automation to be controlled by that setting.  See
[Bug 1524673](https://bugzilla.mozilla.org/show_bug.cgi?id=1524673)
for some of the motivation.  This agrees with Chrome, and should
generalize to future remote control protocols.