DO NOT LAND - Do `org.mozilla.appservices` work "by hand".
This shows how to manually do the megazord substitution that the
`org.mozilla.appservices` plugin does by hand.  It could be factored
into a standalone Gradle file and generalized in a variety of
directions, avoiding the existing plugin, which has proved brittle and
opaque.

There's additional work required order to land this.  I made
`FenixApplication` explicitly depend on the megazord class to verify
that it's doing the right thing; that should be removed.  In fact, we
should make the substitution process for a-s megazord-aware instead
and get rid of the introspection entirely, but that's a separate
problem.