Fix #4372 - Set SitePermissionsDetailsExceptionsFragment toolbar title in onResume
The flow we are currently using involves setting a new title for the containing
Activity when the user navigates to a new Fragment.
This happened for SitePermissionsDetailsExceptionsFragment in it's onCreate().
Opening a permission (SitePermissionsManageExceptionsPhoneFeatureFragment) will
set a new title in it's onCreate() but going back from this new Fragment will
not create again a new SitePermissionsDetailsExceptionsFragment.
As a workaround I moved the call to set Activity's title in onResume().