Fixes #5323: Show a snackbar while browsing data is deleting.
I set the snackbar length indefinite because it should show while the
browsing data is deleting. After that the snackbar is dismissed.
In BaseBrowserFragment I didn't set an anchor for the snackbar because
browserToolbarView is initialized after DefaultBrowserToolbarController
constructor where I pass the snackbar as a parameter (browserToolbarView
needs browserInteractor who needs browserToolbarController, so I
couldn't change the initialization order). The only solution I found was
to pass the snackbar to DefaultBrowserToolbarController via a setter but
this would bypass our current architecture.