For #4690 - Prevent multiple Collections menus and throttle menu clicks
Previously it was possible for the Collections PopupWindow to be displayed
multiple times when using Talkback, situation previously resolved for other
such menus in ticket #3584 that now modified the code to close the menu if one
is already opened.
Also we need to throttle the menu clicks so that the user will not be able to
click on it multiple times - if clicking fast enough it is possible for a second
click to trigger the new dismiss behavior.