Closes #4976: Reuse Fennec launcher entry in manifest.
With this patch Fennec users, who upgrade to Fenix, will not lose Firefox launcher icons they
have added to their home screen.

A side effect of this change is that the launcher entry for existing Fenix users may disappear
from the home screen for existing Firefox Preview users and they will have to re-add it.
Since the Preview audience is more accepting of breaking changes and since that group is
way smaller than the millions of Fennec users, I think that this is an acceptable breaking
change. I looked into supporting multiple launcher entries in parallel, but that got quite
complex with multiple build types and multiple manifest files. Especially if we decide
to keep a Firefox Preview build around (e.g. org.mozilla.fenix) then it would be good to
unify on a single launcher and do that as early as possible.