Fix #8631 - Persist the same activity for http/s schemes after migration
Most of the times setting an application as default does not restrict it's
activities so that only that initial one can handle a specific Intent.
If that activity is renamed/deleted or it's intent filter is now declared for
another, normally, the same application will remain set as default.

IRL we saw that Huawei imposes strict restrictions regarding the default apps
settings, such that any changes in preferred activities would result in
dropping the current default app to let the user pick again a default one.

To protect against such edgecases and keep the default app status previously set
for Fennec we'll keep the same Fennec component declared in Fenix's manifest
as the activity set to handle http/s schemes.