Bug 1619795: Use toolchain-style fetches (and `toolchain-artifact` attribute).
The mozilla-central ffmpeg toolchain (used in Bug 1613479/#8780) has an
artifact that does not start with `public/build/` (it starts with just
`public/`). To easily support that, we want to set the `toolchain-artfiact`
attribute for that task.

We switched `index_search` to be a job-type in
[taskgraph](https://hg.mozilla.org/ci/taskgraph/rev/932b4cf48cc785c31b03f0b9f2cba654bc67bc00)
which pass through `attributes` from the task definition (and is also more
future proof for adding fenix-specific toolchain tasks). This updates taskgraph
and the consumers of the current toolchain.