For #7726 - Support displaying consecutive Snackbars
This flow does not seem supported by the underlying "material.snackbar"
library. When we try to show a new Snackbar the library will update the
properties of the allready on-screen one. While still dismissing it because we
are trying to show a new one.

To properly support our usecase we need to queue the call to show a new
Snackbar and effectively wait until the previous Snackbar is dismissed and
removed from the screen to only then show a new one.