Revert "Bug 1652979 - Stop producing fennecNightly, fenixNightly and fenixBeta (#12225)"
This reverts commit 7e7d69cb8e4514c86c4677b3b18193c5cb66f884.

This caused the nightly FNPRMS tests to fail because the
`fennec-nightly` variant no longer exists. We could change the variant
but then the migration code is not covered. The least time consuming
solution for the FE perf team is to revert the change and atomically
implement the long term solution rather than making an intermediate
change now and another change after the long term solution arrives,
which would skew our dashboards (that we don't have the ability to
annotate) with multiple incomparable builds.