For #162 - add feature build flags
This allows to specify flags at build time that will allow to
enable/disable specific features from fenix. The first use-case for this
is to produce a variant of fenix without proprietary components such as
gms.

For instance, passing the "-Porg.mozilla.fenix.enableFeatureXYZ=false"
will disable the XYZ feature from the build.