For #12099 - Use bookmark keywords in browser searches
For the moment users are not able to view/add/modify bookmark keywords in Fenix
but they can in Firefox running on other platforms.
After syncing the bookmarks Fenix users will be able to use the same already
set keywords in Fenix and have the same search experience with included support
for query placeholders.