For #15543 - Expand / collapse tabs tray depending on all tabs
Now check all tabs - private or normal and decide whether to show the tabs tray
as expanded or collapsed depending on the highest number of tabs in any
category.
Also added support for the grid based tabs tray - will show tabs tray as
expanded if there are more than 3 or more tabs (normal or private) open
(this meaning two rows in the grid layout).