Update bug numbers to URLs for activation ping
Earlier versions of Glean supported bug numbers of the metadata, but this causes
ambiguity, particularly for Fenix where issues exist in both Bugzilla and
Github. This just updates the Bugzilla numbers to full URLs.

We made the use of numbers an error sometime ago, but neglected to enforce this
on "pings" in addition to "metrics".  A future version of `glean_parser` [1] will make
this an error and this will set Fenix up to be ready for that.

[1] https://github.com/mozilla/glean_parser/pull/262