Use a smaller request code for WebAuthnFeature
We can't see the private API that we interact with on the OS, but after
some internal investigation it appears that there might be an upper
limit to the request code we can use.

For now, let's try a value similar to that use in the GVE code to see
our requests are failing because of that.