For #18142: Changed the way the clipboard verifies the URL
The bug indicates that when we have a super huge string, we get terrible jank when clicking on the URL bar. This is because ClibBoardHandler.kt calls into WebURLFinder.kt which has a regex that ends up doing catastrophic backtracking. The behavior can be avoided using the URLStringUtils. However, we do lose a tiny bit of functionality since we don't find all possible URLs as WebURLFinder did