Bug 1609191 - Disable hardware acceleration on devices affected by Adreno shader compilation crash
On various Adreno devices we have experienced crashes in glLinkProgram
due to a suspected driver bug. There appears to be a race condition
between the gecko renderer thread compiling shaders used to render web
content, and the Android UI render thread compiling shaders used to
render the Fenix UI. Disabling hardware acceleration avoids this crash
as shader compilation will only occur on the gecko renderer thread.

Hardware acceleration must be disabled globally in
AndroidManifest.xml.  We can then opt in to hardware acceleration at
runtime on a per-window basis on suitable devices. We do this by
creating an ActivityLifecycleCallbacks which enables hardware
acceleration for the window whenever an activity is created.

We use a hardcoded list of known-buggy device boards, and only enable
hardware acceleration if the device is not in that list. The list was
chosen by selecting the most common boards that this crash occurs
on (covering >90% of crashes in the past month). Some crashes may
still occur on boards not in this list, however the numbers will be
low and crashes infrequent enough that this should not matter.
Likewise, some devices with these boards may be running driver
versions unaffected by this bug, but it should not matter if hardware
acceleration is disabled anyway.