For #21192 - New FenixTheme to style Compositions
Starting with adding the complete list of Firefox colors from which custom
color palettes for light/dark/private mode are built and made available to
child Composable trees.

In an effort to keep the UI consistent these new color palettes do not allow
overriding individual properties though Composables have access to the entire
list of colors and can decide to use another as an exception.