For #22155 - Inactive tabs count telemetry
A quantity probe in the metrics ping means we'll loose the granularity events
provided but it will be easier to extract the values.

For reporting whether the inactive tabs feature is enabled or not we already
have the "preferences.inactive_tabs_enabled" probe so I didn't duplicate this.

(cherry picked from commit 207e85e59dba2a9e21487166137a6770eaa1b0eb)

# Conflicts:
#	app/metrics.yaml
#	app/src/main/java/org/mozilla/fenix/FenixApplication.kt