Prevent default browser marketing from waking user
This will fix #20887.

`IMPORTANCE_DEFAULT` is documented to make an audible alert, and without any logic in here to restrict the time of day at which the notification is sent, that's not reasonable behavior. The audible alert has been observed in the wild to wake up users in the middle of the night and annoy them (see #20887).

This sets `IMPORTANCE_LOW` instead, which is as close as you can get to `IMPORTANCE_DEFAULT` without making sound. This should make sure the marketing message is more well-received.