Fix BrowserRobot.goToHomescreen to wait for home screen to be rendered
It was previously just waiting for the application to be idle, which
is not a reliable check in case the navigation to home is triggered
on completion of an async operation i.e., the app / main thread may
be idle, but the navigation to home hasn't completed yet.

goToHomeScreen further needs to wait for the browser fragment's
home button to exist before clicking it.