For #12717 - Add search support to bookmarks
Allows for an in-memory search of all bookmarks flattened and then the default
folders (like Bookmarks Toolbar, etc) removed.
Search will only start and the current bookmarks list will be updated to
the queried results after the user first enters a valid query (not blank).
Search will end when the SearchView is closed. (Deleting all query characters
will not end the search mode).