[Uplift] For #22870: Renew metrics set to expire in February (#23496)
* For #22870: Renew/remove metrics set to expire in February (#23453)

* For #22870: Renew/remove metrics set to expire in February

* Add data review and add notification owner for never expire metrics

* Close #22956: Define "tags" for metrics in the app repository

Co-authored-by: William Lachance <wlach@protonmail.com>

Define "tags" for metrics in the app repository

Using the new "tags" feature in Glean, we can remove most of the
manual process/pain of defining tags in a seperate Glean annotations
repository (see: https://github.com/mozilla-mobile/fenix/wiki/Add-a-Glean-Annotation-for-an-event)
and just put them beside the metrics themselves. This should make it
much easier to keep this metadata up to date, since it can be added
in the same pull request as the instrumentation itself.

To keep the list of tags up to date with the Fenix issue labels, a new
script `update-glean-tags.py` has been added to the repository. It
should not need to be run often.

This pull request uses a scraped version of the defined tags in the
Glean annotations repository. After it lands, we can remove those and
just use the tagging information here as the main source of information.

Co-authored-by: Roger Yang <royang@mozilla.com>
(cherry picked from commit 08790a09a692e71bcddbc3d27cce63b844383041)