mergify: allow complete-push check as an alternative to complete-pr in queue conditions
PRs created by github actions don't get a pull-request taskcluster graph
scheduled (because that is only done for PRs created by collaborators).
However, if the PR is created in the main fenix repo, typically because
it comes from one of our internal github workflows, then we get a push
graph, which is sufficient to run the required checks.

This change lets mergify consider a green `complete-push` task as
sufficient to queue a PR (in addition to all the other conditions, e.g.
wrt review), which will allow PRs created by the
update-nimbus-experiments workflow to be queued by mergify instead of
having to be merged by a repo admin.