For #26511: Skip waiting for unavailable wallpapers
Previously we were waiting until the AppState is updated with the wallpaper
previously chosen by the user (option persisted in SharedPreferences).
It is possible though that the previously chosen wallpaper has expired (or for
other reasons it is unavailable) in which case the AppState will be updated
with the default wallpaper while in SharedPreferences we still keep the name of
unavailable wallpapers.