For #26511: Prevent stale wallpaper view layout change listeners
In the scenario of setting the wallpaper from both the observer and the
fragment there would be two layout listeners set for scaling the bitmap with
only the first being used immediately and the second being left to observe
future changes as when the keyboard is shown moment at which it would trigger
an unexpected wallpaper scaling.
By avoiding the apply the wallpaper from the observer and only applying it from
HomeFragment there will be only one layout change listener and so the above
issue will also be avoided.