For #26884: Disable CFRs in tests without affecting the main app
Interacting directly with app settings and not cleaning up after tests has the
potential to modify the app behavior.
We can avoid that by ensuring we always use the FeatureSettingsHelper from the
activity rule.

Searched with a simple "[a-zA-Z]*CFR[a-zA-Z]* =" regex and there should be no
other cases where a CFR is disabled permanently from tests.