For #28093 - Try to show Nimbus messages after the messages are loaded
Currently the `Evaluate` action to try showing Nimbus messages is dispatched
from `MessagingFeature` when the app's `onStart` method is called - when the
application is visible.
The `Restore` action which would load the available messages is dispatched
based on a callback that although is set when the app is opened it was seen in
manual tests that it could fire after app's `onStart` and as such the loaded
messages would not be tried to be shown.

Adding a new call to evaluate messages after we know they are loaded will
ensure they will be shown as soon as possible.

The race condition between loading messages and trying to show them is resolved
in version 109 of the app (current app version) by switching from a callback
system for knowing when the Nimbus configuration is done to having all needed
operations run on `Main` in application's `onCreate` making the data available
to when later `MessagingFeature` tries to show the next available message.